refactor/dashboard-layout: introduce DashboardLayout and migrate existing LiveViews to use it instead of Layouts.app
