+++ AGENTS.md
@@ -43,10 +43,11 @@ Keep this root file lean: only project-specific governance, commands, and critic
 
 ## Phoenix v1.8
 
-- LiveView templates must start with `<Layouts.app flash={@flash} ...>` and pass `current_scope` when required.
-- `Layouts` is already aliased through `my_app_web.ex`; do not add redundant aliases.
+- LiveView templates must start with `<DashboardLayout.dashboard flash={@flash} ...>` and pass `current_scope` when required.
+- `DashboardLayout` is aliased in `controlkeel_web.ex`; do not add redundant aliases.
+- The root layout is owned by `ControlKeelWeb.Layouts` (`components/layouts/root.html.heex`); `put_root_layout` uses `{ControlKeelWeb.Layouts, :root}`. `DashboardLayout` holds the in-page dashboard layout only.
 - Fix missing `current_scope` by moving routes into the right `live_session` and passing the assign through layout boundaries.
-- `<.flash_group>` belongs only in `layouts.ex`.
+- `<.flash_group>` belongs only in `dashboard_layout.ex`.
 - Use `<.icon name="hero-x-mark" class="w-5 h-5"/>`; do not call Heroicons modules directly.
 - Use imported `<.input>` for form inputs. If overriding `class`, fully style the input because defaults are not inherited.
 - Router scopes already include aliases; do not add duplicate route aliases.
