+++ lib/controlkeel_web.ex
@@ -82,7 +82,7 @@ defmodule ControlKeelWeb do
 
       # Common modules used in templates
       alias Phoenix.LiveView.JS
-      alias ControlKeelWeb.Layouts
+      alias ControlKeelWeb.DashboardLayout
       alias ControlKeelWeb.ObservabilityLayout
       alias ControlKeelWeb.ObservabilitySessionLayout
